+++ example/FamilyHubs.Example/Program.cs
@@ -1,11 +1,15 @@
-using Microsoft.Extensions.Configuration;
+using FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options;
+using FamilyHubs.SharedKernel.Razor.Health;
+using Microsoft.Extensions.Options;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddRazorPages();
 
-builder.Services.AddFamilyHubs(builder.Configuration);
+builder.Services
+    .AddFamilyHubs(builder.Configuration)
+    .AddFamilyHubsHealthChecks(builder.Configuration);
 
 var app = builder.Build();
 
@@ -26,5 +30,7 @@ app.UseRouting();
 app.UseAuthorization();
 
 app.MapRazorPages();
+
+app.MapFamilyHubsHealthChecks();
 
 app.Run();